requirements.txt
streamlit==1.45.1
torch==2.7.0
torchvision==0.22.0
torchcam==0.4.0
Pillow==11.2.1
matplotlib==3.10.3